엔티티 매핑 테이블 생성 및 디렉토리 구분
