+++ backend/src/main/java/com/myong/backend/domain/entity/business/PayStatus.java
@@ -1,4 +1,4 @@
-package com.myong.backend.domain.entity;
+package com.myong.backend.domain.entity.business;
 
 public enum PayStatus {
     INCOMPLETE,COMPLETE // 미완료, 완료
