+++ backend/src/main/java/com/myong/backend/domain/entity/business/PaymentMethod.java
@@ -1,4 +1,4 @@
-package com.myong.backend.domain.entity;
+package com.myong.backend.domain.entity.business;
 
 public enum PaymentMethod {
     CASH,CARD; // 현금, 카드
