+++ backend/src/main/java/com/myong/backend/domain/entity/business/ReservationStatus.java
@@ -1,4 +1,4 @@
-package com.myong.backend.domain.entity;
+package com.myong.backend.domain.entity.business;
 
 public enum ReservationStatus {
     WAIT,CANCEL,SUCCESS; // 대기, 취소, 성공
