+++ backend/src/main/java/com/myong/backend/domain/entity/designer/JobApplication.java
@@ -1,5 +1,6 @@
-package com.myong.backend.domain.entity;
+package com.myong.backend.domain.entity.designer;
 
+import com.myong.backend.domain.entity.shop.JobPost;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
