+++ backend/src/main/java/com/myong/backend/domain/entity/shop/Menu.java
@@ -1,5 +1,6 @@
-package com.myong.backend.domain.entity;
+package com.myong.backend.domain.entity.shop;
 
+import com.myong.backend.domain.entity.designer.Designer;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
