+++ backend/src/main/java/com/myong/backend/domain/entity/shop/Work.java
@@ -1,4 +1,4 @@
-package com.myong.backend.domain.entity;
+package com.myong.backend.domain.entity.shop;
 
 public enum Work {
     FULLTIME, CONTACT //정규직, 계약직
