+++ backend/src/main/java/com/myong/backend/domain/entity/user/CouponStatus.java
@@ -1,4 +1,4 @@
-package com.myong.backend.domain.entity;
+package com.myong.backend.domain.entity.user;
 
 public enum CouponStatus {
     NOTUSE, USE; // 미사용, 사용
