+++ backend/src/main/java/com/myong/backend/domain/entity/user/CouponType.java
@@ -1,4 +1,4 @@
-package com.myong.backend.domain.entity;
+package com.myong.backend.domain.entity.user;
 
 public enum CouponType {
     FIXED, PERCENT // 고정금액 할인, 퍼센트 할인
