+++ backend/src/main/java/com/myong/backend/domain/entity/usershop/MessageType.java
@@ -1,4 +1,4 @@
-package com.myong.backend.domain.entity;
+package com.myong.backend.domain.entity.usershop;
 
 public enum MessageType {
     TEXT,IMAGE,FILE // 텍스트, 이미지, 파일
